requirements.txt
flask==3.1.*
gunicorn==23.*
line-bot-sdk==3.*
google-genai>=1.0.0
Pillow>=10.0.0
